1.1 Updates
My good friend Waldek Mastykarz made a couple of updates to show
different colors for tasks depending on who the task is assigned to as
well as some style changes per his stile of coding. Additionally he
changed the task update to update both the start date and due date when
a task is dragged and dropped to keep the duration the same.